% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_path.R
\name{audio_path}
\alias{audio_path}
\title{Get path to an audio file in gRappa}
\usage{
audio_path(file)
}
\arguments{
\item{file}{Name of the audio file}
}
\value{
A character string containing the file path
}
\description{
This function returns the full path to an audio file stored in the package.
}
