% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grappa_banner.R
\name{grappa_banner}
\alias{grappa_banner}
\title{gRappa scrolling banner}
\usage{
grappa_banner(
  speed = 8,
  size = 70,
  text_color = "orange",
  bg_color = "black",
  position = c("top", "center", "bottom")
)
}
\arguments{
\item{speed}{Duration of one full scrolling cycle in seconds.
Smaller values make the banner move faster.}

\item{size}{Font size in pixels.}

\item{text_color}{Color of the text.}

\item{bg_color}{Background color of the page.}

\item{position}{Vertical position of the banner: "top", "center", or "bottom".}
}
\description{
Opens a browser window with a continuously scrolling slogan.
}
