% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_audio.R
\name{list_audio}
\alias{list_audio}
\title{List available audio files in gRappa}
\usage{
list_audio()
}
\value{
A character vector with audio file names
}
\description{
This function lists all audio files stored in the package.
}
