% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available dataset files in gRappa}
\usage{
list_datasets()
}
\value{
A character vector with dataset file names
}
\description{
This function lists all dataset files stored in the package.
}
