% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_icons.R
\name{list_icons}
\alias{list_icons}
\title{List available icon files in gRappa}
\usage{
list_icons()
}
\value{
A character vector with icon file names
}
\description{
This function lists all icon files stored in the package.
}
