% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_media.R
\name{list_media}
\alias{list_media}
\title{List all media files in gRappa}
\usage{
list_media()
}
\value{
A named list containing audio and video files
}
\description{
This function lists all audio and video files stored in the package.
}
