% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_video.R
\name{list_video}
\alias{list_video}
\title{List available video files in gRappa}
\usage{
list_video()
}
\value{
A character vector with video file names
}
\description{
This function lists all video files stored in the package.
}
