% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_grappa_id.R
\name{next_grappa_id}
\alias{next_grappa_id}
\title{Get the next available grappa ID}
\usage{
next_grappa_id()
}
\value{
Character string containing the next grappa id.
}
\description{
Reads the current gRappa dataset and returns the next available
identifier in the sequence (e.g. "grp_000007").
}
