% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_audio.R
\name{play_audio}
\alias{play_audio}
\title{Open an audio file from gRappa}
\usage{
play_audio(file)
}
\arguments{
\item{file}{Name of the audio file}
}
\value{
Invisibly returns the file path
}
\description{
This function opens an audio file stored in the package
using the default application of the operating system.
}
