% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_media.R
\name{play_media}
\alias{play_media}
\title{Play a media file from gRappa}
\usage{
play_media(file)
}
\arguments{
\item{file}{Name of the media file}
}
\value{
Invisibly returns the file name
}
\description{
This function detects whether a file is an audio or a video
stored in the package and opens it with the appropriate function.
}
