% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_video.R
\name{play_video}
\alias{play_video}
\title{Open a video file from gRappa}
\usage{
play_video(file)
}
\arguments{
\item{file}{Name of the video file}
}
\value{
Invisibly returns the file path
}
\description{
This function opens a video file stored in the package
using the default application of the operating system.
}
