% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_grappa.R
\name{random_grappa}
\alias{random_grappa}
\title{Play a random media file from gRappa}
\usage{
random_grappa()
}
\value{
Invisibly returns the name of the selected file
}
\description{
This function randomly selects one available audio or video file
stored in the package and opens it.
}
