% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rima_grappa.R
\name{rima_grappa}
\alias{rima_grappa}
\title{Words that rhyme with "grappa"}
\usage{
rima_grappa(n_letters = 4, exclude_self = FALSE)
}
\arguments{
\item{n_letters}{Number of final letters used to define the rhyme.}

\item{exclude_self}{Logical. If TRUE, removes "grappa" from the output.}
}
\value{
A character vector of Italian words rhyming with "grappa".
}
\description{
Returns Italian words that rhyme with "grappa".
}
