% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rime.R
\name{rime}
\alias{rime}
\title{Find Italian rhymes}
\usage{
rime(word, n_letters = 4, exclude_self = FALSE)
}
\arguments{
\item{word}{A word to rhyme.}

\item{n_letters}{Number of final letters used to define the rhyme.}

\item{exclude_self}{Logical. If TRUE, removes the input word from the output.}
}
\value{
A character vector of rhyming Italian words.
}
\description{
Finds Italian words that rhyme with a given word,
using the last n letters of the word.
}
