% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_icons.R
\name{show_all_icons}
\alias{show_all_icons}
\title{Show all icons in the gRappa package}
\usage{
show_all_icons(ncol = 3)
}
\arguments{
\item{ncol}{Number of columns in the icon gallery.}
}
\value{
Invisibly returns the vector of icon file names.
}
\description{
This function displays all PNG icons stored in the
\code{inst/extdata/icons} directory of the package.
}
