% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_icon.R
\name{show_icon}
\alias{show_icon}
\title{Show an icon from the gRappa package}
\usage{
show_icon(name)
}
\arguments{
\item{name}{Character. Name of the icon file (e.g. "grappa_glass.png").}
}
\value{
The image object (invisibly).
}
\description{
This function loads and displays a PNG icon stored in
the \code{inst/extdata/icons} directory of the package.
}
