% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slogan_grappa.R
\name{slogan_grappa}
\alias{slogan_grappa}
\title{Generate silly gRappa slogans}
\usage{
slogan_grappa(n_letters = 4, n_slogans = 1)
}
\arguments{
\item{n_letters}{Number of final letters used to define the rhyme.}

\item{n_slogans}{Number of slogans to generate.}
}
\value{
A character vector of slogans. If \code{n_slogans = 1}, the slogan is also printed.
}
\description{
Generates one or more random slogans using rhymes for "grappa".
}
