% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_grappa.R
\name{stop_grappa}
\alias{stop_grappa}
\title{Stop gRappa playback by closing common media players}
\usage{
stop_grappa()
}
\value{
Invisibly returns TRUE
}
\description{
This function tries to stop media playback by closing common
media players on Linux, macOS, and Windows.
}
