% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_path.R
\name{video_path}
\alias{video_path}
\title{Get path to a video file in gRappa}
\usage{
video_path(file)
}
\arguments{
\item{file}{Name of the video file}
}
\value{
A character string containing the file path
}
\description{
This function returns the full path to a video file stored in the package.
}
